# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers/{providerId}:
    put:
      operationId: upsertProviderConfig
      tags: [providers]
      summary: Create or update provider configuration
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProviderConfigBody"
      responses:
        "200":
          description: Updated provider config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderConfig"
        "201":
          description: Created provider config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderConfig"
    delete:
      operationId: deleteProviderConfig
      tags: [providers]
      summary: Delete provider configuration
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ProviderConfig:
      type: object
      properties:
        id:
          type: integer
        providerId:
          type: string
        apiKey:
          type: string
          nullable: true
          description: Masked as "••••••••" when set
        baseUrl:
          type: string
          nullable: true
        enabled:
          type: boolean
        selectedModel:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - providerId
        - enabled
        - createdAt
        - updatedAt
    UpsertProviderConfigBody:
      type: object
      properties:
        apiKey:
          type: string
          nullable: true
        baseUrl:
          type: string
          nullable: true
        enabled:
          type: boolean
        selectedModel:
          type: string
          nullable: true