# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
  systemPrompt: zod.string().optional(),
  providerId: zod.string().optional(),
  model: zod.string().optional(),
});

/**
 * @summary Generate an image from a text prompt
 */
export const GenerateOpenaiImageBody = zod.object({
  prompt: zod.string(),
  size: zod.enum(["1024x1024", "512x512", "256x256"]).optional(),
});

export const GenerateOpenaiImageResponse = zod.object({
  b64_json: zod.string(),
});

/**
 * @summary List all configured providers
 */
export const ListProviderConfigsResponseItem = zod.object({
  id: zod.number(),
  providerId: zod.string(),
  apiKey: zod.string().nullable().optional(),
  baseUrl: zod.string().nullable().optional(),
  enabled: zod.boolean(),
  selectedModel: zod.string().nullable().optional(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListProviderConfigsResponse = zod.array(ListProviderConfigsResponseItem);

/**
 * @summary Create or update provider configuration
 */
export const UpsertProviderConfigParams = zod.object({
  providerId: zod.string(),
});

export const UpsertProviderConfigBody = zod.object({
  apiKey: zod.string().nullable().optional(),
  baseUrl: zod.string().nullable().optional(),
  enabled: zod.boolean().optional(),
  selectedModel: zod.string().nullable().optional(),
});

export const UpsertProviderConfigResponse = ListProviderConfigsResponseItem;

/**
 * @summary Delete provider configuration
 */
export const DeleteProviderConfigParams = zod.object({
  providerId: zod.string(),
});